requirements.txt
langchain
langchain_core
langchain_ollama
langchain_community
streamlit
python-dotenv
OpenAI
langchain_openai